Move path lock and new test case
